Manual management of publication tables
Add a new configuration option that determines PublicationManager's
ability to modify the publication.

When the DB role used by Electric does not own the table being
subscribed to by a client, it cannot add it to the publication. But
manually adding the table using a more privileged role isn't sufficient,
since Electric will remove it from the publication during the next
schema reconciliation event.

It's the asymmetry between the ability to remove a table from and
lacking the ability to add a table to the publication that has prompted
us to add the manual publishing mode to let the user manually keep the
publication up-to-date with the active shape subscriptions.